* Simple WIP navigation stubbed out. Prepping to have tip screen and other screens * view model/database example removed from tip screen
