+++ .idea/misc.xml
@@ -15,6 +15,7 @@
         <entry key="../../../../layout/compose-model-1659706939231.xml" value="0.5777777777777777" />
         <entry key="../../../../layout/compose-model-1659804274334.xml" value="0.3962962962962963" />
         <entry key="../../../../layout/compose-model-1663043525089.xml" value="0.32592592592592595" />
+        <entry key="../../../../layout/compose-model-1663085469833.xml" value="0.75" />
       </map>
     </option>
   </component>
