+++ app/src/main/java/me/branwin/boojet/data/BoojetDatabase.kt
@@ -6,7 +6,7 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 
-@Database(entities = [Category::class, Entry::class], version = 1)
+@Database(entities = [Category::class, Entry::class], version = 1, exportSchema = false)
 @TypeConverters(Converters::class)
 abstract class BoojetDatabase: RoomDatabase() {
     abstract fun categoryDao(): CategoryDao
