+++ build.gradle
@@ -2,6 +2,7 @@ buildscript {
     ext {
         compose_version = '1.3.0-beta02'
         room_version = '2.4.3'
+        navigation_compose_version = '2.5.2'
     }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
